+++ pyxem/tests/utils/test_cuda_utils.py
@@ -16,7 +16,7 @@
 # You should have received a copy of the GNU General Public License
 # along with pyXem.  If not, see <http://www.gnu.org/licenses/>.
 
-from pyxem.utils import cuda_utils as cutls
+from pyxem.utils import _cuda as cutls
 import dask.array as da
 import numpy as np
 import pytest
