+++ pyxem/utils/expt_utils.py
@@ -30,7 +30,7 @@ from tqdm import tqdm
 from packaging.version import Version
 
 from pyxem.utils.pyfai_utils import get_azimuthal_integrator
-from pyxem.utils.cuda_utils import is_cupy_array
+from pyxem.utils._cuda import is_cupy_array
 from pyxem.utils._deprecated import deprecated
 
 try:
